credit management tool prototype
With the approaching reality of truly mobile payment applications and
products like coin coming out there is a need to leverage user
transaction data on the consumer side. There is an opportunity to
provide consumer tools and mobile applications to make optimized credit
recommendations based on user preferences. I am in the process of
developing a credit rules engine application in python as a project for
one of my master’s program classes. The projected build for the end of
the semester will imitate the virtual interactions of a person’s wallet
with a group of purchases, complete with the basic logic and structure
of the program to run simple credit optimizations.

The point of the application is to work as a platform off of which to
build further personal credit analysis and allow users to achieve an
optimal mix of credit/debit cards used and usage of each given those
registered with your account and user preference (maximizing rewards,
minimizing ultimate interest paid, minimizing % of credit used).